\name{JMbamlss-package}
\alias{JMbamlss-package}
\alias{JMbamlss}

\docType{package}

\title{
  \packageTitle{JMbamlss}
}

\description{
  \packageDescription{JMbamlss}
}

\details{
  The DESCRIPTION file:
  \packageDESCRIPTION{JMbamlss}
  \packageIndices{JMbamlss}
  An overview of how to use the package, including the most important ~~
  ~~ functions ~~
}

\author{
  \packageAuthor{JMbamlss}

  Maintainer: \packageMaintainer{JMbamlss}
}

\references{
  ~~ Literature or other references for background information ~~
}

\keyword{ package }

\seealso{
  ~~ Optional links to other man pages, e.g. ~~
  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

\examples{
  ~~ simple examples of the most important functions ~~
}
