% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfpca_sim.R
\name{MFPCA_cov}
\alias{MFPCA_cov}
\title{Function to calculate the multivariate FPCA for a given covariance matrix and
univariate basis functions}
\usage{
MFPCA_cov(cov, basis_funs)
}
\arguments{
\item{cov}{Covariance matrix of the basis functions coefficients.}

\item{basis_funs}{List with basis functions on each dimension. The basis
functions are funData objects}
}
\description{
Function to calculate the multivariate FPCA for a given covariance matrix and
univariate basis functions
}
