% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MJM_predict.R
\name{MJM_predict}
\alias{MJM_predict}
\title{Prediction of MJM model}
\usage{
MJM_predict(
  object,
  newdata,
  type = c("link", "parameter", "probabilities", "cumhaz"),
  dt,
  id,
  FUN = function(x) {
     mean(x, na.rm = TRUE)
 },
  subdivisions = 7,
  cores = NULL,
  chunks = 1,
  verbose = FALSE,
  ...
)
}
\description{
Note: Writing a predict function is a bit tricky.
For longitudinal prediction, if subject specific predictions are wanted, then
the PCRE terms must be attached to newdata and already evaluated.
If the model uses standardized survival matrices, the different linear 
predictors should be predicted using different data sets.
}
