% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helperfun.R
\name{sim_bamlss_predict}
\alias{sim_bamlss_predict}
\title{Simulation Helper Function - Predict the Results for bamlss-Models}
\usage{
sim_bamlss_predict(m, wd, model_wd, data_wd)
}
\arguments{
\item{m}{Vector containing all the file names of the models.}

\item{wd}{Path to simulations folder.}

\item{model_wd}{Simulation setting folder where the models are saved.}

\item{data_wd}{Simulation data folder.}
}
\description{
This function takes all the models listed in a folder and predicts the fit.
}
