% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simJM.R
\name{simJM}
\alias{simJM}
\title{New Simulation Function Based On FPCs}
\usage{
simJM(
  nsub = 300,
  times = seq(0, 120, 1),
  probmiss = 0.75,
  long_setting = "functional",
  alpha_setting = if (nonlinear) "linear" else "nonlinear",
  dalpha_setting = "zero",
  sigma = 0.3,
  long_df = 6,
  tmax = NULL,
  seed = NULL,
  full = FALSE,
  file = NULL,
  nonlinear = FALSE,
  fac = FALSE,
  efun_type = "Poly",
  ignoreDeg = NULL,
  eval_type = "linear",
  eval_scale = 1
)
}
\description{
Extend the structure given by simJM function in bamlss
Add arguments
efun_type = "Poly", character string specifying the type of eigenfcts (eFun)
ignoreDeg = NULL, vector of numerics specifying the ignored degrees (eFun)
eval_type = "linear", character string specifying the type of eigenvals (eVal)
eval_scale = 1, numeric to scale the eigenvalues
}
